Get the contract for the source and not the destination
There was an issue w/ when the contract for the destination, a mirror, was being sent
into the `getRetrievalPointer` it was trying to send a `RENEW` command to the source
instead of the destination, as the contract didn't have an hd_key. The `RENEW` command
failed at the farmer with an error that the contract can not be changed to a
different farmer.