Improve logging and errors
Improves the ability to identify the source of timeout issues when trying to get storage offers in the Storj network, by making the logs searchable by frame and shard hash id, and logging specific details about the failure. This also changes the error type to ServiceUnavailable which is more fitting, and which can eventually be solved by having overflow farmers staged in the event that storage offers are not available.