Give default values for responseTime and reputation if not already set
This covers the case that a contact is added via Kad which does not have these default values set. Because the contact will need to be updated with `spaceAvailable:true` to start receiving data, we can also verify that responseTime and reputation have default values so that the farmer will be selected to store data.