Fix issue with initial mirror generation
There was an issue where duplicate mirror events could be triggered
for each side of the exchange report, once for the farmer and another
for the client. The replication/mirror is triggered by a successfull
client report.

There was another issue where a mirror would be attempted from/to
the same farmer. There has been a check added to make sure that the
source and destination are not the same.